Move T2369 and fix it for Haskell2010
