+++ tests/tests.cabal
@@ -1944,7 +1944,6 @@ executable unit-pos-3
                     , T2096
                     , T2235
                     , T2237
---                     , T2369
                     , T385
                     , T531
                     , T595a
@@ -2279,6 +2278,7 @@ executable basic-pos
                     , ReflExt08B
                     , ShadowLogicSymbols
                     , SkipDerived00
+                    , T2369
 
     ghc-options:      -fplugin=LiquidHaskell -fkeep-going -O0
     if flag(measure-timings)
